+++ src/main/java/com/travelland/domain/Trip.java
@@ -1,5 +1,6 @@
 package com.travelland.domain;
 
+import com.travelland.domain.member.Member;
 import com.travelland.dto.TripDto;
 import com.travelland.dto.TripDto.Create;
 import jakarta.persistence.*;
