+++ src/main/java/com/travelland/domain/member/RefreshToken.java
@@ -1,4 +1,4 @@
-package com.travelland.domain;
+package com.travelland.domain.member;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
