+++ src/main/java/com/travelland/global/security/JwtAuthorizationFilter.java
@@ -1,7 +1,7 @@
 package com.travelland.global.security;
 
-import com.travelland.domain.Member;
-import com.travelland.domain.RefreshToken;
+import com.travelland.domain.member.Member;
+import com.travelland.domain.member.RefreshToken;
 import com.travelland.global.exception.CustomException;
 import com.travelland.global.exception.ErrorCode;
 import com.travelland.global.jwt.JwtUtil;
