+++ src/main/java/com/travelland/repository/plan/querydsl/CustomPlanLikeRepository.java
@@ -1,6 +1,6 @@
 package com.travelland.repository.plan.querydsl;
 
-import com.travelland.domain.Member;
+import com.travelland.domain.member.Member;
 import com.travelland.domain.plan.PlanLike;
 
 import java.util.List;
