+++ src/main/java/com/travelland/repository/plan/querydsl/CustomPlanScrapRepository.java
@@ -1,7 +1,6 @@
 package com.travelland.repository.plan.querydsl;
 
-import com.travelland.domain.Member;
-import com.travelland.domain.plan.PlanLike;
+import com.travelland.domain.member.Member;
 import com.travelland.domain.plan.PlanScrap;
 
 import java.util.List;
