+++ src/main/java/com/travelland/repository/trip/querydsl/CustomTripLikeRepository.java
@@ -1,6 +1,6 @@
 package com.travelland.repository.trip.querydsl;
 
-import com.travelland.domain.Member;
+import com.travelland.domain.member.Member;
 import com.travelland.domain.TripLike;
 
 import java.util.List;
