+++ src/main/java/com/travelland/repository/trip/querydsl/CustomTripRepositoryV2.java
@@ -1,6 +1,6 @@
 package com.travelland.repository.trip.querydsl;
 
-import com.travelland.domain.Member;
+import com.travelland.domain.member.Member;
 import com.travelland.domain.Trip;
 
 import java.util.List;
