+++ src/main/java/com/travelland/service/plan/PlanScrapService.java
@@ -1,6 +1,6 @@
 package com.travelland.service.plan;
 
-import com.travelland.domain.Member;
+import com.travelland.domain.member.Member;
 import com.travelland.domain.plan.Plan;
 import com.travelland.domain.plan.PlanScrap;
 import com.travelland.dto.PlanDto;
