+++ src/main/java/com/travelland/service/trip/TripLikeService.java
@@ -1,6 +1,7 @@
 package com.travelland.service.trip;
 
 import com.travelland.domain.*;
+import com.travelland.domain.member.Member;
 import com.travelland.dto.TripDto;
 import com.travelland.global.exception.CustomException;
 import com.travelland.global.exception.ErrorCode;
