+++ src/main/java/com/travelland/service/trip/TripScrapService.java
@@ -1,6 +1,6 @@
 package com.travelland.service.trip;
 
-import com.travelland.domain.Member;
+import com.travelland.domain.member.Member;
 import com.travelland.domain.Trip;
 import com.travelland.domain.TripScrap;
 import com.travelland.dto.TripDto;
